PHP-117- Added changes after feedback to the Delete command, JsonShoppingListRepo, and ShoppingListRepoInterface

+++ src/Persistance/Json/JsonShoppingListRepository.php
@@ -100,6 +100,10 @@ class JsonShoppingListRepository implements ShoppingListRepositoryInterface
     {
         $filename = $this->storeAs($slug);
 
+        if (!$this->exists($slug)) {
+            throw new ShoppingListNotFoundException();
+        }
+
         $this->files->unlink($filename);
     }
